+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## (unreleased)
 
 - Adds support for using custom implied relationship strategies.
+- Adds support for "include relationship==*" (#68).
 - Fixes #69 (hierarchical identifiers not working at the top level inside a deployment environment).
 
 ## 1.12.0 (30th June 2021)